+++ ksyun/client/kce/v20180314/models.go
@@ -308,7 +308,7 @@ type DescribeTagResponse struct {
 		TagName       *string `json:"TagName" name:"TagName"`
 		ImageId       *string `json:"ImageId" name:"ImageId"`
 		CreateTime    *string `json:"CreateTime" name:"CreateTime"`
-		SizeByte      *int    `json:"SizeByte" name:"SizeByte"`
+		SizeByte      *int64  `json:"SizeByte" name:"SizeByte"`
 		Author        *string `json:"Author" name:"Author"`
 		DockerVersion *string `json:"DockerVersion" name:"DockerVersion"`
 		Architecture  *string `json:"Architecture" name:"Architecture"`
